change: Moved skc_main to .
