--- compiler/skc_main/tests/sk/enum.sk
@@ -1,11 +1,0 @@
-a = Some<Int>.new(1)
-b = None
-f = fn(x: Maybe<Int>) { x }
-f(a)
-f(b)
-unless a.value == 1; puts "ng Some#value"; end
-
-o = Ok<Int, Error>.new(0)
-e = Fail<Int, Error>.new(Error.new("fail"))
-
-puts "ok"
