--- compiler/skc_main/tests/sk/float.sk
@@ -1,15 +1,0 @@
-class Helper
-  def self.eq(x: Float, y: Float) -> Bool
-    (x-y).abs < 0.000001
-  end
-end
-
-unless Helper.eq(1.0 + 2.0, 3.0) then puts "ng 1" end
-unless Helper.eq(3.0 - 2.0, 1.0) then puts "ng 2" end
-unless Helper.eq(3.0 * 2.0, 6.0) then puts "ng 3" end
-unless Helper.eq(3.0 / 2.0, 1.5) then puts "ng 4" end
-unless Helper.eq(3.0 % 2,   1.0) then puts "ng 5" end
-
-# TODO: unless -3**2 == -9 then puts "ng -3**2" end
-
-puts "ok"
