--- compiler/skc_main/tests/sk/generics.sk
@@ -1,21 +1,0 @@
-class A
-  def self.foo(a: Array<Int>) -> Int
-    a[0]
-  end
-end
-if A.foo([99]) != 99; puts "ng 1"; end
-
-a = Array<Array<Int>>.new
-unless a.length == 0; puts "ng 2"; end
-
-# Type parameter substitution
-class Foo<T>
-  def bar<W>(f: Fn1<T, W>) -> Int
-    foo = Foo<W>.new
-    foo.bar<Int>{|arg: W| 0} if false
-    99
-  end
-end
-unless Foo<String>.new.bar<Int>{|arg: String| 0} == 99; puts "ng Type parameter substitution"; end
-
-puts "ok"
