--- compiler/skc_main/tests/sk/inheritance.sk
@@ -1,58 +1,0 @@
-# Inherit methods
-class Base1
-  def foo -> Int
-    1
-  end
-end
-
-class Sub1 : Base1
-end
-
-sub1 = Sub1.new
-unless sub1.foo == 1 then puts "ng 1" end
-
-# Inherit initializer
-class Base2
-  def initialize
-    @a = 9
-  end
-end
-
-class Mid2 : Base2
-end
-
-class Sub2 : Mid2
-end
-
-sub2 = Sub2.new
-unless sub2.a == 9 then puts "ng 2" end
-
-# Inherit ivars
-class Base3
-  def initialize(i: Int)
-    @a = i
-  end
-end
-
-class Sub3 : Base3
-  def initialize
-    @a = 2
-    @b = 2.0
-  end
-end
-
-sub3 = Sub3.new
-unless sub3.a == 2 then puts "ng 3" end
-unless sub3.b == 2.0 then puts "ng 4" end
-
-## Type compatibility
-#class Base4
-#  def self.foo(x: Base4); end
-#end
-#
-#class Sub4 : Base4; end
-#
-#sub4 = Sub4.new
-#Base4.foo(sub4)
-
-puts "ok"
