--- compiler/skc_main/tests/sk/ivar.sk
@@ -1,11 +1,0 @@
-class A
-  def initialize
-    var @i = 1
-  end
-end
-a = A.new
-unless a.i == 1 then puts "ng 1" end
-a.i = 2
-unless a.i == 2 then puts "ng 2" end
-
-puts "ok"
