--- compiler/skc_main/tests/sk/logical_operator.sk
@@ -1,12 +1,0 @@
-if false then puts "ng 1" end
-if !true then puts "ng 2.5" end
-if !(true and true) then puts "ng 3" end
-if  (true and false) then puts "ng 4" end
-if  (false and true) then puts "ng 5" end
-if  (false and false) then puts "ng 6" end
-if !(true or true) then puts "ng 11" end
-if !(true or false) then puts "ng 12" end
-if !(false or true) then puts "ng 13" end
-if  (false or false) then puts "ng 14" end
-
-puts "ok"
