--- compiler/skc_main/tests/sk/nested_class.sk
@@ -1,14 +1,0 @@
-class A
-  class B; end
-  class C : B
-    def foo(b: B) -> Int
-      1
-    end
-  end
-  # TODO: If B is here, indexing C fails as it finds B#initialize
-  #class B; end
-end
-b = A::B.new
-unless A::C.new.foo(b) == 1 then puts "ng 1" end
-
-puts "ok"
