--- compiler/skc_main/tests/sk/relational_operator.sk
@@ -1,15 +1,0 @@
-unless 0 < 1; puts "ng 1"; end
-if     0 < 0; puts "ng 2"; end
-unless 0 <= 1; puts "ng 3"; end
-unless 0 <= 0; puts "ng 4"; end
-unless 1 > 0; puts "ng 5"; end
-if     0 > 0; puts "ng 6"; end
-unless 1 >= 0; puts "ng 7"; end
-unless 0 >= 0; puts "ng 8"; end
-
-unless 0 == 0; puts "ng 9"; end
-if     0 == 1; puts "ng 10"; end
-unless 0 != 1; puts "ng 11"; end
-if     0 != 0; puts "ng 12"; end
-
-puts "ok"
