--- compiler/skc_main/tests/sk/string.sk
@@ -1,31 +1,0 @@
-var a = Array<String>.new
-
-# interpolation
-x = 1; y = 2
-unless "x=#{x}, y=#{y}" == "x=1, y=2"; puts "interpolation1: fail"; end
-b = [1,2,3]; c = [4,5]
-unless "b=\{b}, c=\{c}" == "b=[1, 2, 3], c=[4, 5]"; puts "interpolation2: fail"; end
-
-# split
-a = "a<>bc<>d".split("<>")
-unless a.length == 3; puts "split1: bad length"; end
-unless a[0] == "a"; puts "split1: fail a"; end
-unless a[1] == "bc"; puts "split1: fail b"; end
-unless a[2] == "d"; puts "split1: fail c"; end
-
-a = "abc".split("<>")
-unless a.length == 1; puts "split2: bad length"; end
-unless a[0] == "abc"; puts "split2: fail abc"; end
-
-a = "".split("<>")
-unless a.length == 0; puts "split3: bad length"; end
-
-a = "abc".split("")
-unless a.length == 1; puts "split4: bad length"; end
-unless a[0] == "abc"; puts "split4: fail abc"; end
-
-a = "abc<>".split("<>")
-unless a.length == 1; puts "split5: bad length"; end
-unless a[0] == "abc"; puts "split5: fail abc"; end
-
-puts "ok"
