--- compiler/skc_main/tests/sk/type_system.sk
@@ -1,13 +1,0 @@
-# `(String) -> Int` conforms to `(T) -> Void` (#197)
-[""].each do |s: String|
-  0
-end
-
-class A
-  # `Never` conforms to `Int`
-  def self.todo -> Int
-    panic "todo"
-  end
-end
-
-puts "ok"
