--- compiler/skc_main/tests/sk/xxx-eq.sk
@@ -1,23 +1,0 @@
-var a = 1
-
-a += 2
-if a != 3 then puts "ng +=" end
-a -= 2
-if a != 1 then puts "ng -=" end
-a *= 60
-if a != 60 then puts "ng *=" end
-a /= 5
-if a != 12 then puts "ng /=" end
-a %= 5
-if a != 2 then puts "ng !=" end
-a <<= 4
-if a != 32 then puts "ng <<=" end
-a >>= 1
-if a != 16 then puts "ng >>=" end
-a -= 1
-a &= 9
-if a != 9 then puts "ng &=" end
-a |= 15
-if a != 15 then puts "ng |=" end
-
-puts "ok"
